requirements.txt
groq
langchain
langchain-groq
langchain-community
chromadb
sentence-transformers
python-dotenv
google-generativeai